refactoring base
