--- info.py
@@ -1,6 +1,0 @@
-KEY = "" # your bot api key
-ADMINS = [] # list of telegram users ids with special rights
-DB_USER = "" # your username for database access
-DB_PASS = "" # your password for database access
-VIP = [] # list of telegram users ids with VIP status
-WEATHER_KEY = "" # api key for accuweather.com
