+++ pyproject.toml
@@ -1,0 +1,36 @@
+[tool.poetry]
+name = "schedule-bot"
+version = "0.1.0"
+description = ""
+authors = ["okumuramura"]
+license = "MIT"
+
+[tool.poetry.dependencies]
+python = "^3.8"
+SQLAlchemy = "^1.4.32"
+aiogram = "^2.20"
+aiohttp = "^3.8.1"
+aioschedule = "^0.5.2"
+emoji = "^1.7.0"
+mysqlclient = "^2.1.0"
+numpy = "^1.22.4"
+xlrd = "^2.0.1"
+tqdm = "^4.64.0"
+beautifulsoup4 = "^4.11.1"
+
+[tool.poetry.dev-dependencies]
+pytest = "^7.0"
+pytest-cov = "^3.0.0"
+autoflake = "^1.4"
+isort = "^5.10.0"
+black = "^22.1"
+unify = "^0.5"
+flake8 = "^4.0"
+pylint = "^2.12.1"
+mypy = "^0.940"
+pytest-mock = "^3.7.0"
+safety = "^1.10.3"
+
+[build-system]
+requires = ["poetry-core>=1.0.0"]
+build-backend = "poetry.core.masonry.api"
