--- requirements.txt
@@ -1,6 +1,0 @@
-aiogram==2.11.2
-aiohttp==3.7.4.post0
-aioschedule==0.5.2
-emoji==1.5.0
-SQLAlchemy==1.3.23
-mysqlclient==2.0.3
