+++ schedule_bot/logger.conf
@@ -1,0 +1,22 @@
+[loggers]
+keys=root
+
+[handlers]
+keys=consoleHandler
+
+[formatters]
+keys=mainFormatter
+
+[logger_root]
+level=INFO
+handlers=consoleHandler
+
+[handler_consoleHandler]
+class=StreamHandler
+level=DEBUG
+formatter=mainFormatter
+args=(sys.stdout,)
+
+[formatter_mainFormatter]
+format=%(asctime)s - %(name)s - %(levelname)s: %(message)s
+datefmt=%Y-%m-%d %H:%M:%S
