+++ setup.cfg
@@ -1,0 +1,90 @@
+[tool:pytest]
+addopts =
+	--cov=schedule_bot
+	--cov-fail-under=90
+python_files = test_*.py
+python_classes =
+    *Test
+    Test*
+python_functions = test_*
+
+[flake8]
+enable-extensions = G
+exclude = .git, .venv
+ignore =
+    A003 ; 'id' is a python builtin, consider renaming the class attribute
+    E203 ; whitespace before ':"
+    E501 ; line too long
+    W503 ; line break before binary operator
+    W504 ; Line break occurred after a binary operator
+    N802 ; function name 'CONFIG' should be lowercase
+    N803 ; argument name 'CONFIG' should be lowercase
+max-complexity = 15
+show-source = true
+
+[coverage:run]
+omit =
+    tests/*
+    **/__main__.py
+    *site-packages*
+branch = True
+
+[coverage:report]
+show_missing = True
+skip_covered = True
+fail_under = 90
+exclude_lines =
+    pragma: no cover
+    def __repr__
+    def __str__
+
+[isort]
+balanced_wrapping = true
+default_section = THIRDPARTY
+include_trailing_comma = true
+known_first_party = tests,schedule_bot
+line_length = 80
+multi_line_output = 3
+not_skip = __init__.py
+
+[pylint]
+generated-members = responses.*
+good-names = i,j,k,e,x,_,pk,id
+max-module-lines = 300
+output-format = colorized
+ignored-classes=
+    contextlib.closing,
+    sqlalchemy.orm.scoping.scoped_session,
+disable=
+    C0103, ; Constant name "api" doesn"t conform to UPPER_CASE naming style (invalid-name)
+    C0111, ; Missing module docstring (missing-docstring)
+    C0330, ; Wrong hanging indentation before block (add 4 spaces)
+    E0213, ; Method should have "self" as first argument (no-self-argument) - N805 for flake8
+    R0201, ; Method could be a function (no-self-use)
+    R0901, ; Too many ancestors (m/n) (too-many-ancestors)
+    R0903, ; Too few public methods (m/n) (too-few-public-methods)
+    R0914, ; Too many local variables (m/n) (too-many-locals)
+    W0511, ; TODO needed? (fixme)
+    E0611, ; No name '<name>' in module '<module>' (no-name-in-module)
+# Maximum number of nested code blocks for function / method
+max-nested-blocks=3
+# Maximum number of branch for function / method body
+max-branches=15
+# Maximum number of statements in function / method body
+max-statements=30
+# Maximum number of attributes for a class (see R0902).
+max-attributes=10
+# Maximum number of public methods for a class (see R0904).
+max-public-methods=10
+
+[mypy]
+check_untyped_defs = true
+disallow_any_generics = true
+disallow_incomplete_defs = true
+disallow_untyped_defs = true
+ignore_missing_imports = true
+no_implicit_optional = true
+
+[mypy-tests.*]
+disallow_untyped_defs = false
+disallow_incomplete_defs = false
